Require email verification before adding tasks
Email verification is now required before a user can add a new
task. There is a misleading alert saying that it was added,
but it will not be added to either front or backend without
verification.

Also, changed a use of == to === in html because type coercian,
I guess. Thanks, Obamascript.